move layouts to core module
