+++ src/app/app-routing.module.ts
@@ -7,8 +7,8 @@ import { RegisterPageComponent } from './core/auth/pages/register-page/register-
 import { AuthGuard } from './core/guards/auth.guard';
 import { GuestGuard } from './core/guards/guest.guard';
 import { DashboardPageComponent } from './modules/dashboard/pages/dashboard-page/dashboard-page.component';
-import { AuthLayoutComponent } from './shared/layouts/auth-layout/auth-layout.component';
-import { MainLayoutComponent } from './shared/layouts/main-layout/main-layout.component';
+import { AuthLayoutComponent } from './core/layouts/auth-layout/auth-layout.component';
+import { MainLayoutComponent } from './core/layouts/main-layout/main-layout.component';
 import { PageNotFoundComponent } from './shared/pages/page-not-found/page-not-found.component';
 
 const ROUTES: Routes = [
