+++ src/app/app.module.ts
@@ -6,8 +6,8 @@ import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { CoreModule } from './core/core.module';
 import { MainAppModule } from './modules/main-app.module';
-import { AuthLayoutComponent } from './shared/layouts/auth-layout/auth-layout.component';
-import { MainLayoutComponent } from './shared/layouts/main-layout/main-layout.component';
+import { AuthLayoutComponent } from './core/layouts/auth-layout/auth-layout.component';
+import { MainLayoutComponent } from './core/layouts/main-layout/main-layout.component';
 import { PageNotFoundComponent } from './shared/pages/page-not-found/page-not-found.component';
 
 @NgModule({
