+++ src/app/core/layouts/auth-layout/auth-layout.component.spec.ts
